Phase 4 Wave A: move the review desk into surfaces/review/
Third git-mv batch of the core/+surfaces/ scaffolding (admin was
#1070, party #1072). ReviewDesk and its supporting modules — reviewApi,
reviewChecklist, reviewClientContext, reviewTypes — plus the proposal
review cards (ProposalReviewCard, ProposalReviewPanel, ProposalDiffView,
shared between the review desk and Creator Studio's own proposals tab)
move to apps/web/src/surfaces/review/.

proposalsApi.ts stays put: an exhaustive grep found it imported by
ProposalsPage.tsx, ProposeComposer.tsx, RemixPanel.tsx, and
ContributionsSetting.tsx well beyond review — a shared community/
proposals primitive, not review-specific. Same call as leaving
mp/protocol.ts alone in the party batch.

Cross-surface imports repointed at the new paths: App.tsx's ReviewDesk
lazy route, CreatorStudioView.tsx's and surfaces/admin/AdminConsole.tsx's
ProposalReviewPanel import, NavHeader.tsx's and assessmentExportApi.ts's
reviewApi import, and surfaces/admin/AdminAssessmentResolve.tsx/
AdminAssessmentsPanel.(tsx|test.tsx)'s reviewApi/reviewChecklist/
reviewClientContext imports (already pointed at '../../' from the admin
move; now '../review/').

Verified: tsc clean, full apps/web suite (205 files / 1783 tests)
green, repo lint exits 0, production build succeeds (ReviewDesk still
its own ~10.5KB lazy chunk), Playwright check of /review renders the
reviewer-gate correctly.
