+++ apps/web/src/App.tsx
@@ -53,7 +53,7 @@ import { useDocumentTitle } from './useDocumentTitle.js';
 // editor, the remix panel), the review desk, or party's multiplayer client.
 const AdminConsole = lazy(() => import('./surfaces/admin/AdminConsole.js').then((m) => ({ default: m.AdminConsole })));
 const CreatorStudioView = lazy(() => import('./CreatorStudioView.js').then((m) => ({ default: m.CreatorStudioView })));
-const ReviewDesk = lazy(() => import('./ReviewDesk.js').then((m) => ({ default: m.ReviewDesk })));
+const ReviewDesk = lazy(() => import('./surfaces/review/ReviewDesk.js').then((m) => ({ default: m.ReviewDesk })));
 const PartyPage = lazy(() => import('./PartyPage.js').then((m) => ({ default: m.PartyPage })));
 
 /**
