+++ apps/web/src/NavHeader.tsx
@@ -8,7 +8,7 @@ import { Mascot } from './Mascot.js';
 import { NotificationBell } from './NotificationBell.js';
 import { PixelIcon } from './PixelIcon.js';
 import { fetchAdminSummary } from './surfaces/admin/adminApi.js';
-import { fetchReviewStatus } from './reviewApi.js';
+import { fetchReviewStatus } from './surfaces/review/reviewApi.js';
 import { creatorPath } from './router.js';
 import { usePageScrolling } from './usePageScrolling.js';
