+++ apps/web/src/surfaces/admin/AdminAssessmentsPanel.tsx
@@ -9,8 +9,8 @@ import {
 } from './adminApi.js';
 import { AssessmentResolutionSummary, AssessmentResolveForm } from './AdminAssessmentResolve.js';
 import { fetchAllAdminAssessments, type AdminAssessmentsExport } from '../../assessmentExportApi.js';
-import { formatAssessmentChecklist } from '../../reviewChecklist.js';
-import { formatAssessmentClientContext } from '../../reviewClientContext.js';
+import { formatAssessmentChecklist } from '../review/reviewChecklist.js';
+import { formatAssessmentClientContext } from '../review/reviewClientContext.js';
 
 export function AdminAssessmentsPanel() {
   const [data, setData] = useState<AdminAssessmentsExport | null>(null);
