+++ eslint-rules/comment-prose-baseline.json
@@ -585,6 +585,10 @@
     "apps/web/src/surfaces/admin/SuggestionsPanel.tsx": 113,
     "apps/web/src/surfaces/admin/WaitlistPanel.test.tsx": 11,
     "apps/web/src/surfaces/admin/WaitlistPanel.tsx": 164,
+    "apps/web/src/surfaces/review/ProposalDiffView.tsx": 133,
+    "apps/web/src/surfaces/review/ProposalReviewCard.test.tsx": 96,
+    "apps/web/src/surfaces/review/ProposalReviewCard.tsx": 272,
+    "apps/web/src/surfaces/review/ProposalReviewPanel.tsx": 184,
     "apps/web/src/telemetry.test.ts": 201,
     "apps/web/src/telemetry.ts": 968,
     "apps/web/src/TelemetryCharts.test.tsx": 26,
