+++ eslint-rules/module-size-baseline.json
@@ -837,6 +837,17 @@
     "apps/web/src/surfaces/admin/SuggestionsPanel.tsx": 146,
     "apps/web/src/surfaces/admin/WaitlistPanel.test.tsx": 197,
     "apps/web/src/surfaces/admin/WaitlistPanel.tsx": 346,
+    "apps/web/src/surfaces/review/ProposalDiffView.tsx": 86,
+    "apps/web/src/surfaces/review/ProposalReviewCard.test.tsx": 117,
+    "apps/web/src/surfaces/review/ProposalReviewCard.tsx": 197,
+    "apps/web/src/surfaces/review/ProposalReviewPanel.tsx": 56,
+    "apps/web/src/surfaces/review/reviewApi.ts": 223,
+    "apps/web/src/surfaces/review/reviewChecklist.ts": 20,
+    "apps/web/src/surfaces/review/reviewClientContext.test.ts": 63,
+    "apps/web/src/surfaces/review/reviewClientContext.ts": 78,
+    "apps/web/src/surfaces/review/ReviewDesk.test.tsx": 279,
+    "apps/web/src/surfaces/review/ReviewDesk.tsx": 612,
+    "apps/web/src/surfaces/review/reviewTypes.ts": 32,
     "apps/web/src/telemetry.test.ts": 285,
     "apps/web/src/telemetry.ts": 273,
     "apps/web/src/TelemetryCharts.test.tsx": 197,
